SDK-31: fix URL generation for component in GH
The problem was that in dev mode, import.meta.url, which is used to
construct the URL to GitHub component, gets passed as is.
For production build, the URI from import.meta.url is escaped.
To cover for both cases, the path used in this component is
normalized (decoded), so that both dev and prod will have the
same value (dev will not change, but it will change for prod),
and only then it is escaped, producing the same result for both envs.